backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
motor==3.6.0
python-jose==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
httpx==0.28.1
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
